# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    post:
      tags: [Shifts]
      operationId: createShift
      summary: Record the result of a completed shift
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShiftRequest'
      responses:
        '201':
          description: Shift recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftRecord'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      tags: [Shifts]
      operationId: getShifts
      summary: Get the authenticated user's shift history
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: List of shifts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftListResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CreateShiftRequest:
      type: object
      required: [scenarioId, won, scoreEarned, timeSecs, ragePeak, managerUsed]
      properties:
        scenarioId:
          type: string
          minLength: 1
        won:
          type: boolean
        scoreEarned:
          type: integer
        timeSecs:
          type: integer
        ragePeak:
          type: integer
        managerUsed:
          type: boolean
    ShiftRecord:
      type: object
      required: [id, userId, scenarioId, won, scoreEarned, timeSecs, ragePeak, managerUsed, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        scenarioId:
          type: string
        won:
          type: boolean
        scoreEarned:
          type: integer
        timeSecs:
          type: integer
        ragePeak:
          type: integer
        managerUsed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ShiftListResponse:
      type: object
      required: [shifts]
      properties:
        shifts:
          type: array
          items:
            $ref: '#/components/schemas/ShiftRecord'