# lib/api-zod/src/generated/types/shiftRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShiftRecord {
  id: number;
  userId: string;
  scenarioId: string;
  won: boolean;
  scoreEarned: number;
  timeSecs: number;
  ragePeak: number;
  managerUsed: boolean;
  createdAt: Date;
}